LPD-61657 Use CMS feature flag: LPD-17564

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/service/impl/ObjectFieldLocalServiceImpl.java
@@ -340,7 +340,9 @@ public class ObjectFieldLocalServiceImpl
 					ObjectFieldUtil.getCounterName(objectField));
 			}
 
-			if (objectField.compareBusinessType(
+			if (FeatureFlagManagerUtil.isEnabled(
+					objectField.getCompanyId(), "LPD-17564") &&
+				objectField.compareBusinessType(
 					ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
 
 				_ploEntryLocalService.deletePLOEntries(
@@ -947,7 +949,9 @@ public class ObjectFieldLocalServiceImpl
 			objectField, objectDefinition, objectFieldBusinessType,
 			objectFieldSettings, null);
 
-		if (objectDefinition.isApproved() &&
+		if (FeatureFlagManagerUtil.isEnabled(
+				objectField.getCompanyId(), "LPD-17564") &&
+			objectDefinition.isApproved() &&
 			Objects.equals(
 				objectField.getBusinessType(),
 				ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
@@ -1231,7 +1235,10 @@ public class ObjectFieldLocalServiceImpl
 				objectField.getBusinessType(),
 				ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
 
-			if (objectDefinition.isApproved()) {
+			if (FeatureFlagManagerUtil.isEnabled(
+					objectField.getCompanyId(), "LPD-17564") &&
+				objectDefinition.isApproved()) {
+
 				_resourceActions.removeModelResource(
 					objectDefinition.getClassName(),
 					ObjectFieldConstants.
@@ -1607,7 +1614,9 @@ public class ObjectFieldLocalServiceImpl
 				newObjectField, objectDefinition, objectFieldBusinessType,
 				objectFieldSettings, oldObjectField);
 
-			if (businessType.equals(
+			if (FeatureFlagManagerUtil.isEnabled(
+					newObjectField.getCompanyId(), "LPD-17564") &&
+				businessType.equals(
 					ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
 
 				_addOrUpdateObjectFieldResourceActionPLOEntries(newObjectField);
